fix: move hooks to hooks folder
