+++ src/module.ts
@@ -1,14 +1,15 @@
 import defu from 'defu'
 import { resolve } from 'pathe'
 import { defineNuxtModule, addPlugin } from '@nuxt/kit'
-import type { Nuxt } from '@nuxt/schema'
 import type { AlgoliaOptions, CrawlerPage } from './types'
-import { createGenerateDoneHook, createPageGenerateHook } from './crawler'
+import { createGenerateDoneHook, createPageGenerateHook } from './hooks/crawler'
 
 export default defineNuxtModule<AlgoliaOptions>({
-  name: '@nuxt-modules/algolia',
-  configKey: 'algolia',
-  setup (options: AlgoliaOptions, nuxt: Nuxt) {
+  meta: {
+    name: '@nuxt-modules/algolia',
+    configKey: 'algolia'
+  },
+  setup (options, nuxt) {
     if (!options.apiKey) {
       throw new Error('Missing `apiKey`')
     }
